Switch to int fields as memory isn't a priority
git-svn-id: https://threeten.svn.sourceforge.net/svnroot/threeten/trunk/jsr-310-ri@143 291d795c-afe8-5c46-8ee5-bf9dd72e1864

+++ src/main/java/javax/time/calendar/DateTimeHM.java
@@ -76,37 +76,17 @@ public final class DateTimeHM
     /**
      * The month of year being represented.
      */
-    private final byte monthOfYear;
+    private final int monthOfYear;
     /**
      * The day of month being represented.
      */
-    private final byte dayOfMonth;
+    private final int dayOfMonth;
     /**
      * The minute of day.
      */
-    private final short minuteOfDay;
+    private final int minuteOfDay;
 
     //-----------------------------------------------------------------------
-//    /**
-//     * Obtains an instance of <code>DateTimeHM</code>.
-//     *
-//     * @param year  the year to represent, from MIN_VALUE + 1 to MAX_VALUE
-//     * @param monthOfYear  the month of year to represent, from 1 (January) to 12 (December)
-//     * @param dayOfMonth  the day of month to represent, from 1 to 31
-//     * @param minuteOfDay  the minute of day to represent, from 0 to 1439
-//     * @return a DateTimeHM instance
-//     * @throws IllegalCalendarFieldValueException if any field is invalid
-//     */
-//    private static DateTimeHM dateTime(int year, int monthOfYear, int dayOfMonth, int minuteOfDay) {
-//        if (ISOChronology.INSTANCE.validateDate(year, monthOfYear, dayOfMonth) == false) {
-//            int[] resolved = CalendricalResolvers.strict().resolveYMD(year, monthOfYear, dayOfMonth);
-//            year = resolved[0];
-//            monthOfYear = resolved[1];
-//            dayOfMonth = resolved[2];
-//        }
-//        return new DateTimeHM(year, monthOfYear, dayOfMonth, minuteOfDay);
-//    }
-
     /**
      * Obtains an instance of <code>DateTimeHM</code>.
      *
@@ -155,9 +135,9 @@ public final class DateTimeHM
      */
     private DateTimeHM(int year, int monthOfYear, int dayOfMonth, int minuteOfDay) {
         this.year = year;
-        this.monthOfYear = (byte) monthOfYear;
-        this.dayOfMonth = (byte) dayOfMonth;
-        this.minuteOfDay = (short) minuteOfDay;
+        this.monthOfYear = monthOfYear;
+        this.dayOfMonth = dayOfMonth;
+        this.minuteOfDay = minuteOfDay;
     }
 
     //-----------------------------------------------------------------------
